requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
--trusted-host pypi.org
--trusted-host files.pythonhosted.org
--trusted-host pypi.python.org

alembic==1.14.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==24.2.0
    # via
    #   frictionless
    #   jsonschema
    #   referencing
azure-core==1.32.0
    # via
    #   azure-identity
    #   azure-keyvault-secrets
azure-identity==1.19.0
    # via -r requirements.in
azure-keyvault-secrets==4.9.0
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
coverage==7.6.8
    # via -r requirements.in
cryptography==44.0.0
    # via
    #   azure-identity
    #   fastapi-azure-auth
    #   msal
    #   pyjwt
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
fastapi[all]==0.115.6
    # via
    #   -r requirements.in
    #   fastapi-azure-auth
fastapi-azure-auth==5.0.1
    # via -r requirements.in
fastapi-cli[standard]==0.0.6
    # via fastapi
frictionless==5.18.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.0
    # via
    #   fastapi
    #   fastapi-azure-auth
humanize==4.11.0
    # via frictionless
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-keyvault-secrets
    #   frictionless
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via
    #   fastapi
    #   frictionless
jsonschema==4.23.0
    # via frictionless
jsonschema-specifications==2024.10.1
    # via jsonschema
mako==1.3.7
    # via alembic
markdown-it-py==3.0.0
    # via rich
marko==2.1.2
    # via frictionless
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
msal==1.31.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
orjson==3.10.12
    # via fastapi
packaging==24.2
    # via pytest
petl==1.7.15
    # via frictionless
pluggy==1.5.0
    # via pytest
portalocker==2.10.1
    # via msal-extensions
pycparser==2.22
    # via cffi
pydantic==2.10.3
    # via
    #   -r requirements.in
    #   fastapi
    #   frictionless
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-extra-types==2.10.0
    # via fastapi
pydantic-settings==2.6.1
    # via
    #   -r requirements.in
    #   fastapi
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.10.1
    # via
    #   fastapi-azure-auth
    #   msal
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.24.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via frictionless
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.19
    # via fastapi
python-slugify==8.0.4
    # via frictionless
pyyaml==6.0.2
    # via
    #   fastapi
    #   frictionless
    #   uvicorn
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   azure-core
    #   frictionless
    #   msal
rfc3986==2.0.0
    # via frictionless
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.11.2
    # via fastapi-cli
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
simpleeval==1.0.3
    # via frictionless
six==1.16.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   alembic
starlette==0.41.3
    # via fastapi
stringcase==1.2.0
    # via frictionless
tabulate==0.9.0
    # via frictionless
text-unidecode==1.3
    # via python-slugify
tomli==2.2.1
    # via pytest
typer==0.15.0
    # via
    #   fastapi-cli
    #   frictionless
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   azure-core
    #   azure-identity
    #   azure-keyvault-secrets
    #   fastapi
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   uvicorn
ujson==5.10.0
    # via fastapi
urllib3==2.2.3
    # via requests
uvicorn[standard]==0.32.1
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
validators==0.34.0
    # via frictionless
watchfiles==1.0.0
    # via uvicorn
websockets==14.1
    # via uvicorn
